% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{caterpillar_plot}
\alias{caterpillar_plot}
\title{caterpillar_plot}
\usage{
caterpillar_plot(obj, ci = 0.9, col_id = TRUE, ...)
}
\arguments{
\item{obj}{...}

\item{ci}{...}

\item{col_id}{...}

\item{...}{Currently not in use}
}
\description{
caterpillar_plot
}
