% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{posterior_summary}
\usage{
posterior_summary(obj, ci = 0.9, as_df = FALSE, digits = 2)
}
\arguments{
\item{obj}{...}

\item{ci}{...}

\item{as_df}{...}

\item{digits}{...}
}
\description{
posterior_summary
}
