% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ranef_summary}
\alias{ranef_summary}
\title{ranef_summary}
\usage{
ranef_summary(obj, ci = 0.9, as_df = FALSE, digits = 2)
}
\arguments{
\item{obj}{...}

\item{ci}{...}

\item{as_df}{...}

\item{digits}{...}
}
\description{
ranef_summary
}
