% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ranef.R
\name{ss_ranef_beta}
\alias{ss_ranef_beta}
\title{ss_ranef_beta}
\usage{
ss_ranef_beta(
  y,
  X,
  unit,
  burnin = 1000,
  iter = 1000,
  chains = 4,
  priors = NULL,
  vars2monitor = c("alpha", "beta", "gamma", "sigma", "tau1", "tau2", "rho", "theta1",
    "theta2")
)
}
\arguments{
\item{y}{...}

\item{X}{...}

\item{unit}{...}

\item{burnin}{...}

\item{iter}{...}

\item{chains}{...}

\item{priors}{...}

\item{vars2monitor}{...}
}
\description{
ss_ranef_beta
}
