% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ranef.R
\name{ss_ranef_mv}
\alias{ss_ranef_mv}
\title{ss_ranef_mv}
\usage{
ss_ranef_mv(
  Y,
  X,
  unit,
  burnin = 1000,
  iter = 1000,
  chains = 4,
  priors = NULL,
  vars2monitor = c("B", "theta", "gamma1", "gamma2", "sigma", "Tau", "rb", "rw")
)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{unit}{...}

\item{burnin}{...}

\item{iter}{...}

\item{chains}{...}

\item{priors}{...}

\item{vars2monitor}{...}
}
\description{
ss_ranef_mv
}
