% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_ranef.R
\name{ss_ranef_alpha}
\alias{ss_ranef_alpha}
\title{ss_ranef_alpha}
\usage{
ss_ranef_alpha(
  y,
  unit,
  burnin = 1000,
  iter = 1000,
  chains = 4,
  priors = NULL,
  vars2monitor = c("alpha", "gamma", "lambda", "sigma", "tau", "theta")
)
}
\description{
ss_ranef_alpha
}
